+++ src/App.tsx
@@ -1,6 +1,7 @@
 import { BrowserRouter, Route, Routes } from "react-router";
 import CargoOrders from "./pages/CargoOrder.tsx";
 import MenuBar from "./components/MenuBar.tsx";
+import NotFound from "./pages/NotFound";
 
 function App() {
 
@@ -10,6 +11,7 @@ function App() {
                 <Routes>
                     <Route element={<MenuBar />}>
                         <Route path="/" element={<CargoOrders />} />
+                        <Route path="*" element={<NotFound />} />
                     </Route>
                 </Routes>
             </BrowserRouter>
